+++ packages/hardhat-zksync-verify/src/constants.ts
@@ -67,3 +67,8 @@ Please try one of the following options:
 
 export const SINGLE_FILE_CODE_FORMAT = 'solidity-single-file';
 export const JSON_INPUT_CODE_FORMAT = 'solidity-standard-json-input';
+
+export const UNSUCCESSFUL_VERIFICATION_MESSAGE = (networkName: string, address: string) =>
+    `Failed to verify contract on network '${networkName}' with address ${address}`;
+
+export const UNSUCCESSFUL_VERIFICATION_ID = -1;